Conflicts:
	src/rdb_protocol/query_language.cc

+++ src/arch/io/disk/aio.hpp
@@ -53,7 +53,7 @@ public:
         bool get_is_write() const { return !get_is_read(); }
         off_t get_offset() const { return this->u.c.offset; }
         size_t get_count() const { return this->u.c.nbytes; }
-        bool get_succeeded() const { return io_result == (int)this->u.c.nbytes; }
+        bool get_succeeded() const { return io_result == static_cast<int64_t>(this->u.c.nbytes); }
         bool get_errno() const { return -io_result; }
 
     private:
